Added packages.
- renamed Main to ZombieGame
- organized classes into packages
- removed faulty constructor from SurvivorParty
- removed faulty constructor from ZombieParty
