+++ ZombieGame/src/main/java/zombiegame/app/battle/Fatality.java
@@ -1,3 +1,5 @@
+package zombiegame.app.battle;
+
 /**
  * Team Project - ZombieGame
  * Concordia St. Paul - CSC 422
@@ -14,7 +16,7 @@
  * Contains the name of the attacker and casualty.
  * Generated by the Party classes and added to BattleInfo under the direction of BattleControl.
  */
-class Fatality {
+public class Fatality {
     private String attacker; // Name of Character that attacked
     private String casualty; // Name of Character killed by attacker
