+++ ZombieGame/src/main/java/zombiegame/app/character/CommonInfected.java
@@ -1,3 +1,5 @@
+package zombiegame.app.character;
+
 /**
  * CommonInfected Subclass for ZombieGame project
  */
