+++ ZombieGame/src/main/java/zombiegame/app/control/BattleControl.java
@@ -1,3 +1,8 @@
+package zombiegame.app.control;
+
+import zombiegame.app.battle.BattleInfo;
+import zombiegame.app.party.Party;
+
 /**
  * Team Project - ZombieGame
  * Concordia St. Paul - CSC 422
