+++ ZombieGame/src/main/java/zombiegame/app/control/GameControl.java
@@ -1,3 +1,10 @@
+package zombiegame.app.control;
+
+
+import zombiegame.app.battle.BattleInfo;
+import zombiegame.app.party.Party;
+import zombiegame.ui.GameView;
+
 /**
  * Team Project - ZombieGame
  * Concordia St. Paul - CSC 422
