+++ ZombieGame/src/main/java/zombiegame/app/party/SurvivorParty.java
@@ -1,3 +1,10 @@
+package zombiegame.app.party;
+
+import zombiegame.app.character.Child;
+import zombiegame.app.character.Soldier;
+import zombiegame.app.character.Teacher;
+import zombiegame.app.character.Character;
+
 /**
  * Creates of Party Object (collection) of the Survivor type
  */
@@ -19,28 +26,6 @@ public class SurvivorParty extends Party{
     // Constructors
     // Default
     public SurvivorParty(){}
-    
-    // Generates a random Survivor party of specified size 
-    // ************* NOTE: Functionality may be moved to BuildControl class if needed ***********
-    public SurvivorParty(int partySize){
-        
-        for(int i=0; i < partySize; i++){
-                
-            // Random int generator 
-            int randomNum = (int)(Math.random() * 100);  // 0 to 99
-
-            // 25% chance of to add soldier (NOTE: change if needed)
-            if(randomNum < 25){
-                this.add(new Soldier());
-            // 25% chance of to add teacher (NOTE: change if needed)
-            } else if (randomNum >= 25 && randomNum < 50){
-                this.add(new Teacher());
-            } else {
-            // 50% chance of to add child (NOTE: change if needed)
-                this.add(new Child());
-            }
-        } 
-    }
 
     @Override
     public int[] getPartyInfo() {
