+++ ZombieGame/src/main/java/zombiegame/app/party/ZombieParty.java
@@ -1,3 +1,9 @@
+package zombiegame.app.party;
+
+import zombiegame.app.character.Tank;
+import zombiegame.app.character.Character;
+import zombiegame.app.character.CommonInfected;
+
 /**
  * Creates of Party Object (collection) of the Survivor type
  */
@@ -18,24 +24,6 @@ public class ZombieParty extends Party{
     // Constructors
     // Default
     public ZombieParty(){}
-    
-    // Generates a random Zombie party of specified size 
-    // ************* NOTE: Functionality may be moved to BuildControl class if needed ***********
-    public ZombieParty(int partySize){
-        
-        for(int i=0; i < partySize; i++){
-            // Random int generator 
-            int randomNum = (int)(Math.random() * 100);  // 0 to 9
-
-            if(randomNum < 20){
-                // 20% chance of to add Tank (NOTE: change if needed)
-                this.add(new Tank());
-            } else {
-                // 80% chance of to add CommonInfected (NOTE: change if needed)
-                this.add(new CommonInfected());
-            }
-        }
-    }
     
     @Override
     public int[] getPartyInfo() {
